requirements.txt
Flask>=2.2,<3.0
supabase>=2.4.0
python-dotenv>=1.0.1
psycopg2-binary
